base home page card and bind methods to build on, handling unsupported cell types
